Handle Workspace ID errors and update feature flag name
Any failure to compute a workspace ID due to missing inputs from the job message (as we have seen in artifact deletion jobs and maintenance jobs) will result in fallback to workspaceID-less behavior which is fine.

The name of the variable controlling the feature has been changed so it can be enabled on only the new agent, without exercising the broken behavior on previous releases.