Fix null ref exceptions on artifact deletion jobs
We have seen the following Listener error in prod while debugging
artifact deletion job issues. On reproducing locally it
turned out this where these jobs fail when *not* using workspace IDs.

Object reference not set to an instance of an object.
   at Microsoft.VisualStudio.Services.Agent.Listener.JobDispatcher.RunAsync(AgentJobRequestMessage message, WorkerDispatcher previousJobDispatch, CancellationToken jobRequestCancellationToken, CancellationToken workerCancelTimeoutKillToken)
   at Microsoft.VisualStudio.Services.Agent.Listener.JobDispatcher.EnsureDispatchFinished(WorkerDispatcher jobDispatch, Boolean cancelRunningJob)

Locally I traced the null ref to the identifier variable.
From the previous code it looks like the intention was to default
to "0", but those defaults were being nulled out by the TryGetValue
calls, so I fixed that default behavior.